Fix change detection relating to vpc_config
As reported in https://github.com/hashicorp/terraform-provider-aws/issues/17385
a change i  detected in vpc_config when there are no changes, this is
caused by an issue in https://github.com/hashicorp/terraform-plugin-sdk/issues/617
The PR to provide a fix is not getting any traction.

On further debuging the issue is caused by the nested elements being of
type set which needs the use of Equal rather than reflect.DeepEqual to
test for differences.

We can work around this bug by testing for changes in the two fields
within vpc_config independantly as when the item passed to HasChanges is
a Set it is tested correctly.